Change rule for Haskell2XmlNew into XmlContent, add Parse rule.
The Haskell2XmlNew class was a temporary staging post on the way to
developing a new XmlContent class, available in HaXml-1.14 onwards.
This patch updates the name, and the rule, to match the latest
developments.  It also adds support for the textual (non-XML) Parse
class, which is an experimental replacement for the Read class, using
monadic parser combinators.  This is also currently defined in
HaXml-1.14.

darcs-hash:20051127105914-6871e-5f29ba081d8d5dde1faed43484790a0067f6178d.gz